// frontend/src/shared/types.ts
export interface components {
    schemas: {
        /** BlogPostOut */
        BlogPostOut: {
            /** Title */
            title: string;
            /** Image */
            image?: string | null;
            /** Paragraphs */
            paragraphs: string[];
            /** Authored By */
            authored_by?: string | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /** Updated At */
            updated_at?: string | null;
            /** Id */
            id: number;
            /** Slug */
            slug: string;
        };
        /** BlogPostsWithAuthors */
        BlogPostsWithAuthors: {
            /** Total Count */
            total_count: number;
            /** Current Page */
            current_page: number;
            /** Page Size */
            page_size: number;
            /** Total Pages */
            total_pages: number;
            /** Items */
            items: components["schemas"]["BlogPostOut"][];
            /**
             * Authors
             * @default []
             */
            authors: string[];
        };
        /** Body_admin_login_admin_login_post */
        Body_admin_login_admin_login_post: {
            /** Email */
            email: string;
            /** Password */
            password: string;
        };
        /** Body_create_blog_post_admin_blogs_new_post */
        Body_create_blog_post_admin_blogs_new_post: {
            /** Title */
            title: string;
            /** Paragraphs */
            paragraphs: string;
            /** Authored By */
            authored_by?: string;
            /**
             * Image
             * Format: binary
             */
            image?: string;
        };
        /** Body_update_blog_post_admin_blogs__post_id__edit_post */
        Body_update_blog_post_admin_blogs__post_id__edit_post: {
            /** Title */
            title: string;
            /** Paragraphs */
            paragraphs: string;
            /** Authored By */
            authored_by?: string;
            /**
             * Image
             * Format: binary
             */
            image?: string;
        };
        /** Body_update_profile_users_me_patch */
        Body_update_profile_users_me_patch: {
            /** Profile Picture */
            profile_picture?: string | null;
        };
        /** FacebookLoginPayload */
        FacebookLoginPayload: {
            /** User */
            user: {
                [key: string]: unknown;
            };
            /** State */
            state?: string | null;
        };
        /** GoogleLoginPayload */
        GoogleLoginPayload: {
            /** Id Token */
            id_token: string;
            /** State */
            state?: string | null;
        };
        /** HTTPValidationError */
        HTTPValidationError: {
            /** Detail */
            detail?: components["schemas"]["ValidationError"][];
        };
        /** PasswordResetConfirm */
        PasswordResetConfirm: {
            /** Token */
            token: string;
            /** New Password */
            new_password: string;
        };
        /** PasswordResetRequest */
        PasswordResetRequest: {
            /**
             * Email
             * Format: email
             */
            email: string;
        };
        /** UserCreate */
        UserCreate: {
            /**
             * Email
             * Format: email
             */
            email: string;
            /** Password */
            password: string;
            /** First Name */
            first_name: string;
            /** Last Name */
            last_name: string;
            /** Redirect To */
            redirect_to?: string | null;
        };
        /** UserLogin */
        UserLogin: {
            /**
             * Email
             * Format: email
             */
            email: string;
            /** Password */
            password: string;
        };
        /** UserRead */
        UserRead: {
            /** Id */
            id: number;
            /**
             * Email
             * Format: email
             */
            email: string;
            /** First Name */
            first_name: string;
            /** Last Name */
            last_name: string;
            /** Profile Picture */
            profile_picture?: string | null;
        };
        /** UserUpdate */
        UserUpdate: {
            /** First Name */
            first_name: string;
            /** Last Name */
            last_name: string;
            /** Profile Picture */
            profile_picture?: string | null;
        };
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}